requirements.txt
# Workana job monitor + Discord webhooks + local LLM bids
playwright>=1.40.0
httpx>=0.25.0
python-dotenv>=1.0.0
pyinstaller>=6.0.0
